rename: OauthStatus -> Type으로 enum 명명 변경
