+++ src/main/java/bit/user/dto/UserCreateRequest.java
@@ -1,7 +1,7 @@
 package bit.user.dto;
 
 import bit.user.domain.User;
-import bit.user.oauth.enums.OauthPlatformStatus;
+import bit.user.oauth.enums.OauthPlatformType;
 import bit.user.oauth.kakao.domain.KakaoUserInfo;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
@@ -25,13 +25,13 @@ public class UserCreateRequest {
     private final String gender;
 
     @Schema(description = "가입 플랫폼", example = "KAKAO", required = true)
-    private final OauthPlatformStatus platform;
+    private final OauthPlatformType platform;
 
     @Schema(description = "가입 일시")
     private final LocalDateTime registerDate;
 
     @Builder
-    public UserCreateRequest(Long id, String email, String nickName, String gender, OauthPlatformStatus platform,
+    public UserCreateRequest(Long id, String email, String nickName, String gender, OauthPlatformType platform,
                              LocalDateTime registerDate) {
         this.id = id;
         this.email = email;
@@ -45,7 +45,7 @@ public class UserCreateRequest {
         return UserCreateRequest.builder()
                 .email(info.getEmail())
                 .nickName(info.getNickname())
-                .platform(OauthPlatformStatus.KAKAO)
+                .platform(OauthPlatformType.KAKAO)
                 .registerDate(info.getConnectedAt())
                 .build();
     }
