+++ src/test/java/bit/user/domain/UserTest.java
@@ -1,6 +1,6 @@
 package bit.user.domain;
 
-import static bit.user.oauth.enums.OauthPlatformStatus.KAKAO;
+import static bit.user.oauth.enums.OauthPlatformType.KAKAO;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import bit.user.dto.UserCreateRequest;
